+++ app/src/main/java/com/merive/securely/activities/LoginActivity.java
@@ -1,10 +1,9 @@
 package com.merive.securely.activities;
 
-import static com.merive.securely.elements.TypingTextView.typingAnimation;
+import static com.merive.securely.components.TypingTextView.typingAnimation;
 
 import android.content.Context;
 import android.content.Intent;
-import android.media.AudioManager;
 import android.os.Bundle;
 import android.os.Handler;
 import android.view.inputmethod.EditorInfo;
@@ -15,7 +14,7 @@ import android.widget.ImageView;
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.merive.securely.R;
-import com.merive.securely.elements.TypingTextView;
+import com.merive.securely.components.TypingTextView;
 import com.merive.securely.preferences.PreferencesManager;
 import com.merive.securely.utils.VibrationManager;
