+++ app/src/main/res/layout/fragment_confirm.xml
@@ -7,7 +7,7 @@
     android:paddingTop="@dimen/padding_12dp"
     android:paddingBottom="@dimen/padding_12dp">
 
-    <com.merive.securely.elements.TypingTextView
+    <com.merive.securely.components.TypingTextView
         android:id="@+id/confirm_title_text"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
