+++ app/src/main/res/layout/fragment_password.xml
@@ -8,7 +8,7 @@
     android:paddingTop="@dimen/padding_12dp"
     android:paddingBottom="@dimen/padding_12dp">
 
-    <com.merive.securely.elements.TypingTextView
+    <com.merive.securely.components.TypingTextView
         android:id="@+id/password_title_text"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
